Add more javadocs and improve test
